parsing: implement gas flow calculation
Calculate the gas flow in m3/min.
