+++ SensateIoT.SmartEnergy.Dsmr.Parser.Common/Logic/SystemClock.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using SensateIoT.SmartEnergy.Dsmr.Parser.Common.Abstract;
 
-namespace SensateIoT.SmartEnergy.Dsmr.Parser.Common.Services
+namespace SensateIoT.SmartEnergy.Dsmr.Parser.Common.Logic
 {
 	public class SystemClock : ISystemClock
 	{
