+++ SensateIoT.SmartEnergy.Dsmr.Parser.Contracts/DTO/GasData.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.Serialization;
+
+namespace SensateIoT.SmartEnergy.Dsmr.Parser.Contracts.DTO
+{
+	public class GasData
+	{
+		
+		[DataMember]
+		public decimal GasConsumption { get; set; }
+		[DataMember]
+		public decimal GasFlow { get; set; }
+	}
+}
